Refactor memory handling and loop in ATTLIST parser
When parsing a duplicate attribute spec in an ATTLIST,
the results are put in a dummy attribute_t structure
while the spec is parsed for validity. There can be
more than one of these in an ATTLIST, but we were only
properly cleaning up the last.

Rewrite the loop so that
* we get rid of all the gotos; instead quit at the
beginning of the loop if we are in error.
* write a short function which clears up the attribute_t.
* call this function in the three appropriate places