Correction to DOM entity decl parsgin
When DOM parser encounters entity declarations, it should not try and
parse their contents immediately. Rather it should wait until
the end of the DTD, since order of entity declaration is not
important.

Also fix but in initial_entity handling for SAX parser, they
are general entities not parameter entities