Basic support for NetCDF's ncml xml format
This XML language is used to encode the metadata
assoceated with a netCDF dataset. This new module
allows writing of the language and automatic
generation of ncml fragments from an existing
dataset. A simple example of use can be found in ncdfcat,
which reads a NetCDF file 'test.nc' and writes
a ncml fragment to 'test.xml'

Much work still to do here.