Allow wxml, wcml and sax to compiler with pgi
pgf90 7.1-4 would not compile wxml or sax modules
because it raised a compile-time type missmatch
when passing null() as an argument to subroutine
that expected a pointer to a derived type. Fix
this by creating a pointer of the relevent type
and setting this to null(). Wrapped in #ifdef PGF90,
which is set up by the configure script.