Change link order to work with pgf90
For reasons that are not clear the order of that
the utils and common libaries are passed to the
linker by FoX-config makes a difference to pgf90
once the compiler bug workaround is applied. This
changes the order and allows wxml, wcml and sax
to work again with the pgi compiler.