Workaround for sed -i non-portability for release
The release scripts and cutdown make target use sed
with inplace file editing (-i), but BSD/OSX and GNU
sed are incopatable in their handling of -i with no
backup extension (BSD needs -i "" and GNU does not
like the empty quoted string after the space). Avoid
this by writing to a temp file and moving this back.