Calculate MD5 and SHA1 digests in release.sh
It's probably worth calculating cryptographic
digests of the .tgz files for release when
we create them in order to allow us to say
if a file has been corrupted between release
time and some later date. This makes the
release script calculate the MD5 and SHA1
digests of each of the .tgz files and
records them in FoX-VERSION-digests
alongside the new archives for release.