Make as_chars take (optional) pair of index args
So you can say as_chars(vs,10,20) for ten of the
characters in the vstr. Note that this is implemeted
as a module procedure because it would be hard to
set the length of the returned fortran string using
the (more natural) 'optional' approach. This makes
it hard to handle Fortran style array slices like
array(:10) or array(10:). Perhapse the solution
would be to use the Perl style slice (negative
counts backwards from the right hand side), but
I'm not sure about this. I think this would allow the
removal of the sanity checks so the function would
be pure.

Also declare vs_len as pure so it can be used in
varable declaration statments.