Make normalize_attribute_text work with vstr's
This function now takes an automatic length fortran
string as input and returns a pointer to a vstr.
This seems to be the cleanest way to do this at the
moment as it is called from places that have both
vstr's and arrays of chars. Internally the function
still uses arrays of chars.