Fix substring overrun in scalartostring
Several of the test cases try to do rts with input
of " abcd " (or similar) and expect to get "abcd"
back. scalartostring wants to return "abcd " but
only has four chars for output data. Under g95
with --enable-debug this leads to a runtime crash
for these tests.

Avoid the porblem by truncating output before trying
to write. The tests now all work and give the same
result with and without the -debug flag set. However,
"abcde" will return "abcd" if the output variable
is four chars long.

Is this OK? Should we set iostat = 1? What about in
the case where we truncate a space (from the tests,
iostat should be 0 in this case).