Add WKML module
WKML is used for generating KML files that can be
read by Google Earth. This is the work of Gen-Tao
Chiang and Toby White.

As of this import there is one known bug (the
contoring fails accross 180 degrees) and a few
items of cleanup needed (we should avoid making
xf public in wxml and need to update .m4 files
in /wkml so that regenerating the .f90 does not
revert recent changes).