Change 'error' module variable from char array to vs
This is used as storage for any errors returned from
the sax_parse call (the error handler stuffs the error
message into the variable just before the sax parser
returns (haveing not fully populated the dom tree).
All that happens then is a check if error is
assoceated, and if it is it if freed (and the
error message currently discarded) and a dom error
generated.