Add more URI tests
It turns out that a couple of the tests listed in rfc2396
don't work. Specifically rebasing with just a query
failes to remove the last segment and a couple of the
'../../../../' type tests don't give expected results
(but I'm not sure what the rule is we are breaking
here). Shouldn't matter for most XML.