Add new funs [Get/Set]Pretty_print, use in wkml
m_wkml_chart uses meaningfull whitespace. For this
to work xf%pretty_print has to be .false. but it is
nice to be able to allow the rest of the document to
be pretty printed. This was done by removing the
private directive from xmlxf_t and twiddling the
logical directly but this approach is not very
nice. Instead add a new function to report the
current setting and a function to change the
setting:

	logical xmlf_GetPretty_print(xf)

returns the current state and:

	call xmlf_SetPretty_print(xf, .false.)

turns off pretty printing. This is now used in
the wkml_chart. Documentation has been added to
DoX.