Build fix for wkml and gfortran - type(color)
For (unknown) reasons gfortran (4.2 - 4.4) would not
compile m_wkml_features.f90, complaining that color
used as an argument to many of the subs clashed with
the same name imported by host assoceation (presumably
the derived type color from m_wkml_color but this was
not clear from the error message and was being renamed
col). Irespective of the cause (a bug in wkml or something
odd in gfortran) this needs fixing and the direct approach
is taken here: rename the derived type color_t.

wkml now builds with gfortran.