New tests of cmlAddParameter for complex values
The PGI compiler seems to have an 'issue' with the complex
double precission version of this generic interface. Specifically
there is a glibc triggered abort triggered from the internal
call to str when the optional fmt argument is missing (but the
tests of the str function don't crash, and they do test the
case triggered from within wcml). This looks like a rather
odd PGI bug specific to complex arguments, a deepish stack
and generic interfaces. The existing wcml tests did not
fail as they use the fmt argument.

Until this is tracked down and worked around / avoided
we can at least test for the problem.