dom: Silence warnigns from m_dom_extras
When exiting in an error condition an output variable
(data) was not set. This results in a warnign from g95
with --enable-debug and gfortran (always). To avoid this
set data="" (for the whole array). This is acheved by
extending the m4 macro TOHW_raise_exception to take a
forth agument (some text to insert before the return
call) and using this to set the value of the returned
(fortran) variable.