utils: Avoid use of uninitialized int in prng seed
The mt_prng is needed for the generation of UUIDs, and
this needs seeding with an integer. We were using an
the contents of two elements of an uninitialized array
on the stack to do this - valgrind complains and I'm not
sure how random this really is (compilers can choose to
set such variables to a particular number for e.g.
debugging). Instead use the output from date_and_time.
Could this lead to problems on systems without a real
time clock? I think we end up overflowing in this case.

The alternitive is to use random_number, but that needs
a call to random_seed (well two calls, one to find the
size of the seed array and one to set it) and that leads
to the same issue.