wxml: Add empty optional argument to cmplx for pgi
The PGI fortran compiler fails with a runtime error for
the various overloded wxml calls with complex arguments
in some cases. The reason is not clear (possibly the
runtime is attempting to deallocate the unused optional
argument to str) and the crash cannot be reproduced in
isolation. However, it has been found that providing
an empty optional argument avoids the crash. So provide
one.