wkml: Simple tests for Contours
Just tests for the three _sp ways of passing the data
in. The irregular cart grid and true irregular grid
produce the same output (as they should) but the
regular grid test produces somewhat different output.
I think this is because the underlying data is not
tied to the long,lat coordinate system but is on the
OS grid (or similar). Hence there is a slight distortion
when forced onto a regular long,lat grid.

I'm not particually sure I like the way the data input
(and test harness) is handled here, need to look for a
better way.