wkml: Fix contour bug for regular grid
The contour rescaling for the regular grid was
forcing the contors to be located around the intersection
of the prime maridian and the equator. The fix is easy
(but there are four options which should give the same
result, unless the grid isn't quite regular).

NB: I think a bug remains when the data crosses the equator
or prime maridian.