wkml: Avoid invalid chars in addCells
Changes in m_wkml_coverage.m4 to better reflect GT's current
version. This seems to work with the example but I don't
have a full test suite for this API yet (and making one
is complicated by the size of the output files) so I'm
not sure if this is a compleate fix.