Check for errors in kml_GetColor
The two specific functions in the GetColor
generic procedure are not pure so we can
check the input arguments and bail out if they
are rubbish. Implement these tests rather then
just return the first color in the X11 list.

For _index the check is that the input value is
between 1 and size(color list) while for _name
the check is that the name was found.

Also remove alternate (default case) return
statment from _name.