build: Avoid errors for paralell make & check
It should be possible to do:

./configure ; make -j 8 check

to have the whole of FoX compiled and tested
concurrently on 8 processors. However, compiling
the tests make use of FoX-config --libs and this
passes location of all the *.a files to the linker.
If any of the libs havn't yet been compiled the linker
will fail at this step and the test will also
fail. This shows up as tests under /utils
failing as e.g. dom_lib has not been built yet.

The work around is to make all the _check targets
depend on all of the main targets so that the *.a
files are in place. i.e. serialize the compilation
and testing steps.