common: Allow errors/warnings to dump core
Add and test a new feature that allows the global
error and warning handelers to become fatal and
cause a stack trace / core dump to be produced.

This is a global setting for all XML documents being
manipulated. Two subroutines take a single logical
argument to turn on (true) and off (false) the feature
for warnings and errors respectivly:

   FoX_set_fatal_warnings
   FoX_set_fatal_errors

and two functions (without arguments) allow the state to
be checked:

   FoX_get_fatal_errors
   FoX_get_fatal_warnings

Both fatal warnings and errors are off by default. This
corresponds to the previous behaviour. All four procedures
must be imported by USEing FoX_common.

Also, add some brief docs for this new feature to DoX.