wkml: Allow headings to be non integers and > 180
According to the schema the heading to a point
style is a kml:angle360Type which can be between
-360 and 360 degrees and is an xsl double data type.
This is not what the google docs say (0 -> 360) or what
was implemented in the code (-180 -> 180).
Change this to follow the schema.

Also allow the heading argument to be an int, or sp or dp
real. The heading must agree with the type of the size
(if there is one). I'm sure this could be coded up much
more consisly.

Update tests - we probably need more now too.