common: Add countrts function
It is potentially useful to find out how many elements
could be returned by the conveniance rts subroutine. This
adds basic support for a countrts function to do this:

  number_of_elements = countrts(string, datatype)

returns the number of elements or zero (if a parse
error was detected).

Datatype is a scalar value used to indicate how the string
should be parsed. This is needed as, e.g. the string
"true false wibble" should return 3 if parsed
to an array of characters, or 0 (an error) if the string
is converted into an array of logicals.

No err value is retuned via an optional argument as this
would prevent the function from becoming pure in future, a
necessesary condition for using the function for initaliazation.

Tests have been integrated with common's test_input
program. Each test also checks that countrts returns the
correct value. Tests where rts would return err=1 are skipped.
Tests which return err=-1 are not errors from the point of
view of countrts, which returns the number of elements needed.
The test code handles these.