common: Don't special case string scalar countrts
The rts subroutine handles the case of scalar string output
differently to arrays of strings in that the output string
is not split into elements (at e.g. whitespace). The first
implementation of countrts handled this by having a different
routine for string scalar and string array datatype arguments
but this is inconsistant. Instead assume the user wants to
know how large an array will be needed if the string is to be
split up.

Hence, we cannot test the string scalar cases in test_input.f90