sax: Allow doctype names to terminate with a ">"
Production 28 of te BNF grammar in XML 1.0 shows that
<!DOCTYPE somename> is well formed and valid if the
root element name is "somename". However, the sax tokenizer
returns "somename>" as the name, which then fails the test
for well formed names in the parser. (On the other hand,
<!DOCTYPE somename > is correctly tokenized.)

This is fixed by allowing a ">" character to end the name, and
setting the following state as ending the tag. We don't need to
use the pushback buffer in this case. However, this is only
needed for the doctype and not the rest of the DTD.