sax: Trap invalid URI errors from open_file
As well as reporting errors by setting iostat,
open_file etc. can signal errors by placing them on
the fx error stack (by calling the add_error subroutine).
One example is generated by parsing an invalid URI. The
user visible open_xml_file sax subroutine does not trap
or report these errors leading to the liklyhood of the
main SAX parser being called after open_xml_file failed
which leads to a runtime crash.

We now trap these errors. Error handling depends on the
presence of the optional iostat argumnt to open_xml_file.
If this is not present and an error is detected, FoX_error
is called with the appropreate error description. This
stops execution and prints the error message. If iostat
is present it is set to SAX_OPEN_ERROR (a default integer
with a value chosen with the hope of not clashing with the
processor-dependent values returned by open). SAX_OPEN_ERROR
is exported via FoX_sax to enable user-side error handling.