fsys: Test for Intel compiler bug
Parsing large files with the SAX or DOM
modules with recent versions of Intel's
compiler results in an access violation
caused by a stack overflow not seen with
other compilers (issue ID DPD200161472):

http://software.intel.com/en-us/forums/showthread.php?t=77847

There is no easy work around, but we can
add a test. This patch adds the test.