wcml: Add tests for cmlAddMolecule
test_cmlAddMolecule_10 checks the optional bondIds
argument works.

test_cmlAddMolecule_11 checks you get an error message
if the atom ref does not exist.

test_cmlAddMolecule_12 checks that if you add a bond
twice you get an error. **NB: this is not yet implemented
and the test FAILS**

test_cmlAddMoleule_13 cheks that the two atom IDs must be
different for a bond and that you get an error if you don't