wcml: Make sure a bond is not added twice.
It is not permitted to have two bonds between the same pair of
atoms. Prevent this by creating a logical array of 'seen' bonds
and check this when we do the other bonding tests. Could probably
move the other tests to this array (items on the diagonal are
not permitted).

The failing wcml test (test_cmlAddMolecule_12) now passes.