fsys/common: Adopt F2008 for specification funcs.
The Fortran 2008 spec adds a new constraint on
how generic procedures can be used in specification
expressions (e.g. the set the length of character
variables or the size of arrays). Specifically,
section 7.1.11 says:

"If a specification expression in the specification
-part of a module ... includes a reference to a generic
entity, that generic entity shall have no specific
procedures defined ... subsequent to the specification
expression."

At least version 7.4.4 of the Cray compiler enforces
this constraint (always, aparantly). So, to enable
FoX to be compiled under a Fortran 2008 compiler,
move functions around in files so that the order is
as needed.

Thanks to Elena Breitmoser (Edinburgh Parallel
Computing Centre) for reporting this issue and
the solution.