wcml: Basic framework for lexical input data dump
We need a way to store input files in CML documents
such that we can recover input. This commit introduces
a first (and incompleate) subroutine to write an input
file (given a file name) to a cml file. Many issues
remain (e.g. stdin input, file that is open etc) but
this (1) introduces the new module, (2) sorts out the
build system, (3) creates a first test case.