wxml/wcml: Add fx%extended_data and use for wcml
In order to implement validity constraints imposed by the
CML schema and compchem convention without breaking
applications that already use FoX_wcml we need a way to
allow FoX_wcml to change behaviour depending on how it
is used. This ability could be useful for other
'writer' modules.

In FoX_wcml add an optional logical input argument to
cmlStartCml which (1) adds two extra namespace declarations
needed for the compchem convention and (2) uses this
namespace to put prefixes on the UUID metadata and convention
values. In future this could also turn on some limited
run-time valididity checking.

In order to keep track of if this is on or off add a
new integer member to the xmlf_t. This is 0 by default
and by stuffing some other number in it the useing
module is expected to change its own behaviour. This
member is accessed by a new function xmlf_GetExtendedData
(which takes a single xmlf_t argument and returns an integer)
and set by a new subroutine, xmlf_SetExtendedData (which takes
two arguments, the xmlf_t type and and integer value to set).

Finally, add a test to show that this actually does something and
update the documentation.