sax: Speedup SAX parser for large text blocks
The SAX parser has to handle arbitary length
strings and in FoX these are mostly stored as
arrays of characters (of length 1). These arrays
typically grow on a character by character basis
by allocating an array one charater longer than
the starting array, copying the data, and
removing the old array. This can be very slow.

This patch introduces a new derived type (varstr)
which contains a character array and a size attirbute.
The array can thus grow in 'chunks' larger thant the
number of characters being added saving many allocate,
copy, deallocate cycles and speeding up the SAX parser.
This is (incidentally) a necessary starting step
for allowing FoX to handle non-ASCII files.

The performance improvment caused by this patch
varies with the size and complexity of the XML
document being processed. Factors of ~500
have been observed for large documents with
long text blocks.