build: Fix FoX-config path after make install
FoX-config reports the location of libs and module
files for users of the FoX. If make install is run
(to install into the --prifix= location) these are
wrong. Update the reported location at install time
such that FoX-config still works if make install is
not run (i.e. FoX is used in-place). The fix is a
bit ugly but seems to work.