build: Avoid crashing for SIESTA build
The final 'touch /.f90 /.F90' command in the configure
script is included directly from configure.ac in order to:

  dnl and to stop over-eager makes trying to recreate f90 files
  dnl when m4 is not gnu m4.

(which is hard to test for unless you happen to have a
non gnu m4). This fails if configure is run from outside
the main FoX directory (e.g. for Siesta when VPATH is used)
and, furthermore, if FoX is configured from within a Makefile
the make will terminate with an error. Wrap the touch command
in a subshell and always report sucess, even if the touch
fails.