wcml: trim elementtype, *id and *ref args
The CML schmea prohibits white space in CML id and
ref attributes (which should not be confused with
the XML id and ref attributes - they have different
syntax and semantics) as well as element names.
While it is not easy to check
whitespace in all cases one easy win is possible,
remove any trailing whitespace. This patch does this
where possible using trim. This cannot be done for
all cases without hacking around with the general
case m4 macros - which are used by other FoX
modules.

Note that if this patch changes the behviour of any
application the application was producing invalid
CML. No test cases needed changing for this to pass.