sax: Fix parsing bug with ']' in PCDATA
Fix bug reported in:
   https://github.com/andreww/fox/issues/23

The SAX parser (which is used by the DOM) looks
out for ']]>' in parsed character data (because
such a string is not permitted - it marks the
end of unparsed character data. In keeping track
of the ']' charcters we forgot to reset the
internal counter when ']a' or ']]a' were found
and then added ']' or ']]' to each additional
character. This is fixed.

Need to audit the code to check the other
cases - is PCDATA case correct for example.
Should add test cases to the sax tests too.
And fix up the external sax & dom tests.