Don't fail if BOM is present
For a UTF8-encoded XML file with a Byte Order Mark and
characters that are also ascii characters, we should be
able to read the file. If the first character is not-
recognisable assume we are dealing with a BOM, skip it,
and carry on. We'll then either read the file OK or we
end up with something that is not well-formed (e.g. because
it is a different encoding).