changed the initialization of IO error codes.
Initializing the IO error codes each time that a xml file
in open for reading may become cumbersome in parallel application
with many processes opening scratch file for tests. In this case
it is preferable to run setup_io just with one node and propagate the
results to all other nodes.
I have redefined setup_io as an interface which refers to 2 module
procedures. setup_io_scracth which is the former  setup_io routine and
a setup_io_with_data where err_code eor_code and eof_code are provided
as arguments with which io_err, io_eor and io_eof are initialized.
I have added a private logical variable io_init that is toggled to true by
setup_io on exit. setup_io checks the io_init variable when starting and
in true returns without no further actions.