SAX Tests for encoding and BOMs
Try reading two files with Byte Order Marks.
test_sax_fsm_1_utf8_bom.in is UTF8 encoded
(but this is not mentioned in the XML header).
We should be able to read this as long as we don't
trip up with the BOM. test_sax_fsm_1_utf16_bom.in
is UTF16 encoded with a BOM and encoding declared
in the XML. We should not be able to read this
(we should get a non-well-formed error).