Move getTextContent_len towards the top of the file
Some versions of gfortran ICE with the function in
the body of the file. Defining it first seems to be
fine and should be valid Fortran 2008.
See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=85138
and https://github.com/andreww/fox/issues/57

Do the actual move in the m4 and rebuild (we comit the
resulting F90 files so users don't need an m4 implementation
to build FoX). At the same time propogate a couple of changes
from the F90 back into other m4 files so we don't loose those
changes.